+++ src/translatex/main.py
@@ -15,7 +15,7 @@ from . import __version__
 from .marker import Marker
 from .preprocessor import Preprocessor
 from .tokenizer import Tokenizer
-from .translator import (Translator, TRANSLATION_SERVICES,
+from .translator import (Translator, TRANSLATION_SERVICE_CLASSES,
                          add_custom_translation_services)
 
 DEFAULT_INTER_FILE_PRE: str = "_"
@@ -69,7 +69,7 @@ def translatex(args: argparse.Namespace) -> None:
             f.write(t.dump_store())
     a = Translator.from_tokenizer(t)
     if not args.dry_run:
-        a.translate(service=TRANSLATION_SERVICES[args.service],
+        a.translate(service=TRANSLATION_SERVICE_CLASSES[args.service](),
                     source_lang=args.src_lang,
                     destination_lang=args.dest_lang)
         if args.stop == "Translator":
@@ -131,7 +131,7 @@ def parse_args(args) -> argparse.Namespace:
     parser.add_argument(
         "-ca", "--custom_api", type=argparse.FileType('r'),
         help="python file that provides a custom translation service class")
-    service_choices = tuple(TRANSLATION_SERVICES.keys()
+    service_choices = tuple(TRANSLATION_SERVICE_CLASSES.keys()
                             ) + ('Custom service...',)
     parser.add_argument(
         "--service",